Fixed some warnings in MSVC. Refactored OpenGL shared context handling to fix an issue where GLEW was initialized too late.

+++ src/Model/ParaxialTexCoordSystem.cpp
@@ -153,9 +153,9 @@ namespace TrenchBroom {
             // WARNING: the texture plane norm is not the rotation axis of the texture (it's always the absolute axis)
             
             // determine the rotation angle from the dot product of the new base axes and the transformed texture axes
-            float cosX = newBaseXAxis.dot(newXAxis);
+            float cosX = static_cast<float>(newBaseXAxis.dot(newXAxis));
             assert(!Math::isnan(cosX));
-            float radX = static_cast<float>(std::acos(cosX));
+            float radX = std::acos(cosX);
             if (crossed(newBaseXAxis, newXAxis).dot(newProjectionAxis) < 0.0)
                 radX *= -1.0f;
